# backend/app/api/v1/projects.py
"""Project records and membership (PROJ-01, PROJ-02, PROJ-07, UI-03)."""

from __future__ import annotations

from typing import Annotated
from uuid import UUID

from fastapi import APIRouter, Query, status

from app.api.deps import ProfScopeDep, ScopeDep, SessionDep
from app.core.pagination import Page
from app.projects import service
from app.projects.schemas import (
    JoinableProjectOut,
    JoinIn,
    MembershipEndIn,
    MembershipIn,
    MembershipOut,
    ProjectIn,
    ProjectOut,
    ProjectPatch,
    ProjectStatus,
)

router = APIRouter(prefix="/projects", tags=["projects"])


@router.get("", summary="List the projects the caller may see")
async def list_projects(
    scope: ScopeDep,
    session: SessionDep,
    project_status: Annotated[ProjectStatus | None, Query(alias="status")] = None,
    limit: Annotated[int | None, Query(ge=1, le=200)] = None,
    cursor: str | None = None,
) -> Page[ProjectOut]:
    return await service.list_projects(
        session, scope, status=project_status, limit=limit, cursor=cursor
    )


@router.get("/joinable", summary="Projects a student may join")
async def list_joinable(scope: ScopeDep, session: SessionDep) -> list[JoinableProjectOut]:
    # Declared above `/{project_id}`: FastAPI matches in registration order, so the UUID path
    # parameter would otherwise claim the literal `joinable` and answer 422 rather than fall
    # through to this route.
    return await service.list_joinable(session, scope)


@router.post("", status_code=status.HTTP_201_CREATED, summary="Create a project")
async def create_project(payload: ProjectIn, scope: ScopeDep, session: SessionDep) -> ProjectOut:
    return await service.create_project(
        session,
        scope,
        title=payload.title,
        description=payload.description,
        stage=payload.stage,
        research_questions=payload.research_questions,
        intended_contributions=payload.intended_contributions,
        start_on=payload.start_on,
        target_on=payload.target_on,
        venue_target=payload.venue_target,
        repo_url=payload.repo_url,
        shared_resources=payload.shared_resources,
    )


@router.get("/{project_id}", summary="Read one project")
async def get_project(project_id: UUID, scope: ScopeDep, session: SessionDep) -> ProjectOut:
    return await service.get_project(session, scope, project_id)


@router.patch("/{project_id}", summary="Update a project")
async def update_project(
    project_id: UUID, payload: ProjectPatch, scope: ScopeDep, session: SessionDep
) -> ProjectOut:
    return await service.update_project(
        session, scope, project_id, **payload.model_dump(exclude_unset=True)
    )


@router.get("/{project_id}/members", summary="List project members")
async def list_members(
    project_id: UUID,
    scope: ScopeDep,
    session: SessionDep,
    include_past: bool = False,
) -> list[MembershipOut]:
    return await service.list_members(session, scope, project_id, include_past=include_past)


@router.post(
    "/{project_id}/members",
    status_code=status.HTTP_201_CREATED,
    summary="Assign a student to the project",
)
async def add_member(
    project_id: UUID, payload: MembershipIn, scope: ProfScopeDep, session: SessionDep
) -> MembershipOut:
    return await service.add_member(
        session,
        scope,
        project_id,
        student_id=payload.student_id,
        responsibility=payload.responsibility,
        joined_on=payload.joined_on,
        planned_allocation=payload.planned_allocation,
    )


@router.post(
    "/{project_id}/join",
    status_code=status.HTTP_201_CREATED,
    summary="Join a project that is open to joining",
)
async def join_project(
    project_id: UUID, payload: JoinIn, scope: ScopeDep, session: SessionDep
) -> MembershipOut:
    return await service.join_project(
        session, scope, project_id, responsibility=payload.responsibility
    )


@router.post(
    "/{project_id}/members/{membership_id}/end",
    summary="End a membership, keeping its history",
)
async def end_membership(
    project_id: UUID,
    membership_id: UUID,
    payload: MembershipEndIn,
    scope: ProfScopeDep,
    session: SessionDep,
) -> MembershipOut:
    return await service.end_membership(
        session, scope, membership_id, left_on=payload.left_on, project_id=project_id
    )