# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JoinIn": {
        "properties": {
          "responsibility": {
            "default": "",
            "title": "Responsibility",
            "type": "string"
          }
        },
        "title": "JoinIn",
        "type": "object"
      },
...
      "LinkProjectIn": {
        "properties": {
          "path_rules": {
            "items": {
              "type": "string"
            },
            "title": "Path Rules",
            "type": "array"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id"
        ],
        "title": "LinkProjectIn",
        "type": "object"
      },
...
      "MembershipEndIn": {
        "properties": {
          "left_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Left On"
          }
        },
        "title": "MembershipEndIn",
        "type": "object"
      },
      "MembershipIn": {
        "properties": {
          "joined_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Joined On"
          },
          "planned_allocation": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Allocation"
          },
          "responsibility": {
            "default": "",
            "title": "Responsibility",
            "type": "string"
          },
          "student_id": {
            "format": "uuid",
            "title": "Student Id",
            "type": "string"
          }
        },
        "required": [
          "student_id"
        ],
        "title": "MembershipIn",
        "type": "object"
      },
...
      "MembershipOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "joined_on": {
            "format": "date",
            "title": "Joined On",
            "type": "string"
          },
          "left_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Left On"
          },
          "origin": {
            "$ref": "#/components/schemas/MembershipOrigin"
          },
          "planned_allocation": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Allocation"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "responsibility": {
            "title": "Responsibility",
            "type": "string"
          },
          "student_id": {
            "format": "uuid",
            "title": "Student Id",
            "type": "string"
          },
          "student_name": {
            "default": "",
            "title": "Student Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "student_id",
          "responsibility",
          "origin",
          "joined_on",
          "created_at"
        ],
        "title": "MembershipOut",
        "type": "object"
      },
...
      "Page_ProjectOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "default": 50,
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[ProjectOut]",
        "type": "object"
      },
...
      "ProjectIn": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "intended_contributions": {
            "items": {
              "type": "string"
            },
            "title": "Intended Contributions",
            "type": "array"
          },
          "repo_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo Url"
          },
          "research_questions": {
            "items": {
              "type": "string"
            },
            "title": "Research Questions",
            "type": "array"
          },
          "shared_resources": {
            "additionalProperties": true,
            "title": "Shared Resources",
            "type": "object"
          },
          "stage": {
            "$ref": "#/components/schemas/ResearchStage"
          },
          "start_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start On"
          },
          "target_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target On"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "venue_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Target"
          }
        },
        "required": [
          "title",
          "stage"
        ],
        "title": "ProjectIn",
        "type": "object"
      },
      "ProjectLinkOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "path_rules": {
            "items": {
              "type": "string"
            },
            "title": "Path Rules",
            "type": "array"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "repository_id": {
            "format": "uuid",
            "title": "Repository Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "repository_id",
          "project_id",
          "path_rules"
        ],
        "title": "ProjectLinkOut",
        "type": "object"
      },
      "ProjectOut": {
        "properties": {
          "ai_restricted": {
            "title": "Ai Restricted",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "intended_contributions": {
            "items": {
              "type": "string"
            },
            "title": "Intended Contributions",
            "type": "array"
          },
          "open_to_join": {
            "title": "Open To Join",
            "type": "boolean"
          },
          "repo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo Url"
          },
          "research_questions": {
            "items": {
              "type": "string"
            },
            "title": "Research Questions",
            "type": "array"
          },
          "shared_resources": {
            "additionalProperties": true,
            "title": "Shared Resources",
            "type": "object"
          },
          "stage": {
            "$ref": "#/components/schemas/ResearchStage"
          },
          "start_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start On"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "target_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target On"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "venue_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Target"
          },
          "viewer_left_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewer Left On"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "title",
          "description",
          "research_questions",
          "intended_contributions",
          "stage",
          "status",
          "shared_resources",
          "ai_restricted",
          "open_to_join",
          "created_at"
        ],
        "title": "ProjectOut",
        "type": "object"
      },
      "ProjectPatch": {
        "properties": {
          "ai_restricted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Restricted"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "intended_contributions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intended Contributions"
          },
          "open_to_join": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open To Join"
          },
          "repo_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo Url"
          },
          "research_questions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Research Questions"
          },
          "shared_resources": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shared Resources"
          },
          "stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "start_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start On"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "target_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target On"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 300,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "venue_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Target"
          }
        },
        "title": "ProjectPatch",
        "type": "object"
      },
      "ProjectStatus": {
        "enum": [
          "proposed",
          "active",
          "paused",
          "completed",
          "archived"
        ],
        "title": "ProjectStatus",
        "type": "string"
      },
...
      "ResearchStage": {
        "description": "The stages the system must support (requirements \u00a71); the rubric adapts to each (PROJ-05).",
        "enum": [
          "literature_review",
          "theory",
          "data_preparation",
          "implementation",
          "experimentation",
          "analysis",
          "writing"
        ],
        "title": "ResearchStage",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Research Management System",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "operationId": "list_projects_api_v1_projects_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProjectStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 200,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProjectOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List the projects the caller may see",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}": {
      "get": {
        "operationId": "get_project_api_v1_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read one project",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "update_project_api_v1_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update a project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/join": {
      "post": {
        "operationId": "join_project_api_v1_projects__project_id__join_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Join a project that is open to joining",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/members": {
      "get": {
        "operationId": "list_members_api_v1_projects__project_id__members_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_past",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Past",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MembershipOut"
                  },
                  "title": "Response List Members Api V1 Projects  Project Id  Members Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List project members",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "add_member_api_v1_projects__project_id__members_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign a student to the project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/members/{membership_id}/end": {
      "post": {
        "operationId": "end_membership_api_v1_projects__project_id__members__membership_id__end_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "membership_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Membership Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipEndIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "End a membership, keeping its history",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/repositories/{repository_id}/projects": {
      "get": {
        "operationId": "list_project_links_api_v1_repositories__repository_id__projects_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectLinkOut"
                  },
                  "title": "Response List Project Links Api V1 Repositories  Repository Id  Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Projects this repository serves",
        "tags": [
          "evidence"
        ]
      },
      "post": {
        "description": "REPO-04: a repository serving several projects leaves unmatched paths unresolved rather\nthan guessing, so the path rules are how attribution gets decided.",
        "operationId": "link_project_api_v1_repositories__repository_id__projects_post",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkProjectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectLinkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Say which project this repository's work belongs to",
        "tags": [
          "evidence"
        ]
      }
    },
...
  }
}